Correctly set parent property on Tk open file/dir dialogs
This is needed for the window manager to prevent the dialog
from going below the main window. Ideally, it should also
be set as modal (on X11), but there doesn't seem to be a solution
to set _NET_WM_STATE_MODAL using Tk.

Rcmdr is likely to do the same change soon.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@30 edb9625f-4e0d-4859-8d74-9fd3b1da38cb