Improve terms dictionary
It's useful to show number of occurrences in addition to the list of words,
and it's a relatively cheap computation. Also create the list before removing
stopwords, so that we can show them and indicate whether they were removed.

Add a "processing" attribute to dtm so that we can find later what operations
where run. It will be useful to add these informations to the report later.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@319 edb9625f-4e0d-4859-8d74-9fd3b1da38cb