Use all potential repositories to install Rstem
CRAN extras provides Rstem binaries for Windows, and it is included
by default in the list returned by getOption("repos"). Also try with
Omegahat in case the package is updated or R-Forge fails. This also
means we can stop hosting Rstem on R-Forge as soon as an Omegahat
or CRAN solution is found.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@429 edb9625f-4e0d-4859-8d74-9fd3b1da38cb