SnowballC: Initial commit
Add a new SnowballC package providing a fully compatible alternative
to Rstem, but based on the libstemmer C library provided by the
Snowball project under the BSD license. The library files are
provided in several subdirectories, but they have all been move
to src/ to eliminate the need to have a Makefile. Only the UTF-8
stemmers have been kept, and useless files have been removed. Other
than that, updating the files to a new version should be easy.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@430 edb9625f-4e0d-4859-8d74-9fd3b1da38cb