Revert "Factiva: Fix parsing IPD for some documents"
This reverts commit 20c409a64faaf9cca4352e8369365dc35893338b.
It turns out that splitting on commas is not a good idea.
Factiva groups several categories into one with commas, but we
have no idea what sub-category a document belongs to: better
leave the user the responsibility of abbreviating the tag if
needed. Solutions may have to be found in the GUI.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-temis@487 edb9625f-4e0d-4859-8d74-9fd3b1da38cb